backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
openai>=1.40.0
pdfplumber
python-dotenv
